youtube_video_upload.py
import google.auth
from googleapiclient.discovery import build
from googleapiclient.errors import HttpError
from googleapiclient.http import MediaFileUpload
import boto3

# First, let's set up the necessary credentials for the YouTube API
# You will need to create a project in the Google Developers Console
# and enable the YouTube Data API v3

credentials = google.auth.credentials.Credentials.from_service_account_info(
    {
        "client_id": "YOUR_CLIENT_ID",
        "client_secret": "YOUR_CLIENT_SECRET",
        "refresh_token": "YOUR_REFRESH_TOKEN",
        "token_uri": "https://accounts.google.com/o/oauth2/token",
        "type": "authorized_user"
    }
)

youtube = build('youtube', 'v3', credentials=credentials)

# Now let's set up the necessary credentials for the AWS S3
# You will need to create an IAM user with the necessary permissions
# to read and list objects in your S3 bucket

s3 = boto3.client(
    's3',
    aws_access_key_id='YOUR_AWS_ACCESS_KEY',
    aws_secret_access_key='YOUR_AWS_SECRET_KEY'
)

# Next, let's create a function to upload the video from S3 to YouTube

def upload_video_to_youtube(video_file_name, video_title, video_description):
    try:
        # Retrieve the video file from S3
        s3_response = s3.get_object(Bucket='YOUR_S3_BUCKET_NAME', Key=video_file_name)
        video_file = s3_response['Body'].read()

        # Create the MediaFileUpload object and set the video file
        media = MediaFileUpload(video_file, mimetype='video/*')

        # Create the video metadata
        video_metadata = {
            'snippet': {
                'title': video_title,
                'description': video_description
            },
            'status': {
                'privacyStatus': 'unlisted'
            }
        }

        # Insert the video
        insert_request = youtube.videos().insert(
            part=','.join(video_metadata.keys()),
            body=video_metadata,
            media_body=media
        )

        insert_request.execute()
        print(f'Video "{video_title}" was successfully uploaded to YouTube.')

    except HttpError as error:
        print(f'An error occurred: {error}')
        raise error
