requirements.txt
google-auth
google-api-python-client
boto3